% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{fetch_latest_submissions}
\alias{fetch_latest_submissions}
\title{Get latest submissions}
\usage{
fetch_latest_submissions()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of recent submissions with columns:
\code{submission_id}, \code{project_title}, \code{submitted_by}, \code{updated_at},
\code{institution}, \code{total_file_size}.
}
\description{
Get latest submissions
}
\examples{
\dontrun{
fetch_latest_submissions()
}
}
