% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas.R
\name{fetch_midas_sources}
\alias{fetch_midas_sources}
\title{Get MIDAS data sources}
\usage{
fetch_midas_sources()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns: \code{key}, \code{label}, \code{has_score},
\code{weight}, \code{row_count}.
}
\description{
List all available gene-disease association databases.
}
\examples{
\dontrun{
fetch_midas_sources()
}
}
