% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netvis.R
\name{fetch_netvis_stats}
\alias{fetch_netvis_stats}
\title{Get NetVis network statistics}
\usage{
fetch_netvis_stats()
}
\value{
A named list with node counts (\code{formula}, \code{herb}, \code{compound},
\code{target}, \code{disease}) and \code{edges} sub-list.
}
\description{
Get NetVis network statistics
}
\examples{
\dontrun{
fetch_netvis_stats()
}
}
