% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{fetch_omics_type_stats}
\alias{fetch_omics_type_stats}
\title{Get omics type statistics}
\usage{
fetch_omics_type_stats()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns: \code{omics_type}, \code{count},
\code{percentage}.
}
\description{
Get omics type statistics
}
\examples{
\dontrun{
fetch_omics_type_stats()
}
}
