% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{fetch_ontology_stats}
\alias{fetch_ontology_stats}
\title{Fetch ontology statistics}
\usage{
fetch_ontology_stats()
}
\value{
A named list with fields: \code{total_entities}, \code{total_level1}--
\code{total_level4}, \code{total_relations}, \code{total_mappings}, \code{categories}.
}
\description{
Fetch ontology statistics
}
\examples{
\donttest{
fetch_ontology_stats()
}
}
