% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptome-hub.R
\name{fetch_transcriptome_stats}
\alias{fetch_transcriptome_stats}
\title{Get Transcriptome Hub statistics}
\usage{
fetch_transcriptome_stats()
}
\value{
A named list with fields: \code{total_datasets}, \code{total_organisms},
\code{total_tcm_entities}, \code{total_analysis_modules}, plus distribution data.
}
\description{
Get Transcriptome Hub statistics
}
\examples{
\dontrun{
fetch_transcriptome_stats()
}
}
