% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingredients.R
\name{get_compound_admet}
\alias{get_compound_admet}
\title{Get ADMET predictions for a compound}
\usage{
get_compound_admet(id)
}
\arguments{
\item{id}{The UniTCM ingredient ID.}
}
\value{
A single-row \code{\link[tibble:tibble]{tibble::tibble()}} with ~90 ADMET columns.
}
\description{
Returns ~90 ADMET endpoint predictions as a single-row wide tibble.
}
\examples{
\dontrun{
get_compound_admet("UNITCM_I00001")
}
}
