% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netvis.R
\name{get_node_detail}
\alias{get_node_detail}
\title{Get node detail}
\usage{
get_node_detail(node_id)
}
\arguments{
\item{node_id}{Node ID.}
}
\value{
A named list with fields: \code{id}, \code{type}, \code{label}, \code{label_cn},
\code{properties}, \code{detail_url}.
}
\description{
Get node detail
}
\examples{
\dontrun{
get_node_detail("H:UNITCM_H001")
}
}
