% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_ontology_by_level}
\alias{get_ontology_by_level}
\title{Get ontology entities by level}
\usage{
get_ontology_by_level(level, parent_id = NULL)
}
\arguments{
\item{level}{Ontology level (integer, 1--4).}

\item{parent_id}{Optional parent entity ID to filter by.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Get ontology entities by level
}
\examples{
\donttest{
get_ontology_by_level(2)
}
}
