% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptome-hub.R
\name{get_transcriptome}
\alias{get_transcriptome}
\title{Get a single transcriptome dataset}
\usage{
get_transcriptome(dataset_id)
}
\arguments{
\item{dataset_id}{The dataset ID (e.g. \code{"TCMtrans00001"}).}
}
\value{
A named list with 35+ fields.
}
\description{
Get a single transcriptome dataset
}
\examples{
\dontrun{
get_transcriptome("TCMtrans00001")
}
}
