% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease-formula.R
\name{list_book_sources}
\alias{list_book_sources}
\title{List book sources}
\usage{
list_book_sources()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{value}, \code{label}, \code{count}.
}
\description{
List book sources
}
\examples{
\donttest{
list_book_sources()
}
}
