% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease-formula.R
\name{list_dosage_forms}
\alias{list_dosage_forms}
\title{List dosage forms}
\usage{
list_dosage_forms()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{value}, \code{label}, \code{count}.
}
\description{
Returns the top 50 dosage forms by frequency.
}
\examples{
\dontrun{
list_dosage_forms()
}
}
