% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{search_ontology_mapping}
\alias{search_ontology_mapping}
\title{Search ontology external mapping}
\usage{
search_ontology_mapping(database, external_id)
}
\arguments{
\item{database}{External database name. Must be one of: \code{"UMLS"},
\code{"SNOMED_CT"}, \code{"ICD11_TM"}, \code{"MeSH"}.}

\item{external_id}{The external identifier to look up.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of matched TCM entities.
}
\description{
Find TCM entities mapped to an external database identifier.
}
\examples{
\donttest{
search_ontology_mapping("MeSH", "D008516")
}
}
