% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_base_url}
\alias{set_base_url}
\title{Set the UniTCM API base URL}
\usage{
set_base_url(url)
}
\arguments{
\item{url}{A character string. The base URL for the UniTCM API,
e.g. \code{"https://UniTCM.cn/api/v1"}.}
}
\value{
Invisible previous URL value.
}
\description{
Set the UniTCM API base URL
}
\examples{
\donttest{
set_base_url("https://UniTCM.cn/api/v1")
}
}
