% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitcm-package.R
\docType{package}
\name{unitcm-package}
\alias{unitcm}
\alias{unitcm-package}
\title{unitcm: R Client for the UniTCM Traditional Chinese Medicine Platform}
\description{
Provides functions to query the UniTCM API \url{https://UniTCM.cn/api/v1/}, covering herb exploration, compound/ADMET data, disease-formula associations, TCM ontology, transcriptomics, and gene-disease analysis (MIDAS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zx122ty/unitcm}
  \item Report bugs at \url{https://github.com/zx122ty/unitcm/issues}
}

}
\author{
\strong{Maintainer}: Xin Zhang \email{designer-why-ninja@duck.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
