% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{as_tidygraph}
\alias{as_tidygraph}
\title{Convert a NetVis graph response to tidygraph}
\usage{
as_tidygraph(graph_response)
}
\arguments{
\item{graph_response}{A list with \verb{$nodes} and \verb{$edges} tibbles, as
returned by \code{\link[=get_neighbors]{get_neighbors()}}, \code{\link[=get_subgraph]{get_subgraph()}}, or \code{\link[=find_path]{find_path()}}.}
}
\value{
A \code{tidygraph::tbl_graph} object.
}
\description{
Convert a NetVis graph response to tidygraph
}
\examples{
\dontrun{
resp <- get_neighbors("H:UNITCM_H001")
tg <- as_tidygraph(resp)
}
}
