% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas.R
\name{autocomplete_disease}
\alias{autocomplete_disease}
\title{Autocomplete disease names (MIDAS)}
\usage{
autocomplete_disease(q)
}
\arguments{
\item{q}{Search query (minimum 2 characters).}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns: \code{disease_name}, \code{disease_id},
\code{gene_count}.
}
\description{
Search for disease names with autocomplete. Query must be at least 2
characters.
}
\examples{
\donttest{
autocomplete_disease("breast")
}
}
