% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{fetch_dataset_facets}
\alias{fetch_dataset_facets}
\title{Get dataset facets}
\usage{
fetch_dataset_facets()
}
\value{
A named list of tibbles for each facet field.
}
\description{
Returns available filter values and their counts for the TCMomics
database.
}
\examples{
\dontrun{
facets <- fetch_dataset_facets()
facets$omics_type
}
}
