% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{fetch_dataset_stats}
\alias{fetch_dataset_stats}
\title{Get TCMomics database statistics}
\usage{
fetch_dataset_stats()
}
\value{
A named list with fields: \code{total_datasets}, \code{total_downloads},
\code{omics_types_count}, \code{unique_organisms}.
}
\description{
Get TCMomics database statistics
}
\examples{
\donttest{
fetch_dataset_stats()
}
}
