% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{fetch_home_stats}
\alias{fetch_home_stats}
\title{Get homepage statistics}
\usage{
fetch_home_stats()
}
\value{
A named list with fields: \code{total_datasets}, \code{total_downloads},
\code{total_file_size}, \code{recent_submissions_count}.
}
\description{
Get homepage statistics
}
\examples{
\dontrun{
fetch_home_stats()
}
}
