% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas.R
\name{fetch_midas_stats}
\alias{fetch_midas_stats}
\title{Get MIDAS statistics}
\usage{
fetch_midas_stats()
}
\value{
A named list with fields: \code{total_associations}, \code{total_genes},
\code{total_diseases}, \code{sources}.
}
\description{
Get MIDAS statistics
}
\examples{
\dontrun{
fetch_midas_stats()
}
}
