% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{fetch_tcm_classification_stats}
\alias{fetch_tcm_classification_stats}
\title{Get TCM classification statistics}
\usage{
fetch_tcm_classification_stats()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns: \code{classification}, \code{count},
\code{percentage}.
}
\description{
Get TCM classification statistics
}
\examples{
\donttest{
fetch_tcm_classification_stats()
}
}
