% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptome-hub.R
\name{fetch_transcriptome_filters}
\alias{fetch_transcriptome_filters}
\title{Get transcriptome filter options}
\usage{
fetch_transcriptome_filters()
}
\value{
A named list of character vectors for each filter field.
}
\description{
Get transcriptome filter options
}
\examples{
\dontrun{
filters <- fetch_transcriptome_filters()
filters$organism
}
}
