% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netvis.R
\name{find_path}
\alias{find_path}
\title{Find shortest path between two nodes}
\usage{
find_path(source, target, max_depth = 4L)
}
\arguments{
\item{source}{Source node ID.}

\item{target}{Target node ID.}

\item{max_depth}{Maximum path depth (max 8, default 4).}
}
\value{
A named list with \verb{$nodes} (\code{\link[tibble:tibble]{tibble::tibble()}}) and \verb{$edges}
(\code{\link[tibble:tibble]{tibble::tibble()}}).
}
\description{
Find shortest path between two nodes
}
\examples{
\dontrun{
find_path("H:UNITCM_H001", "T:TP53")
}
}
