% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptome-hub.R
\name{get_analysis_modules}
\alias{get_analysis_modules}
\title{List available analysis modules for a dataset}
\usage{
get_analysis_modules(dataset_id)
}
\arguments{
\item{dataset_id}{The dataset ID.}
}
\value{
A character vector of available module names.
}
\description{
List available analysis modules for a dataset
}
\examples{
\donttest{
get_analysis_modules("TCMtrans00001")
}
}
