% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netvis.R
\name{get_node_metrics}
\alias{get_node_metrics}
\title{Get node metrics}
\usage{
get_node_metrics(node_id)
}
\arguments{
\item{node_id}{Node ID.}
}
\value{
A named list with fields: \code{node_id}, \code{degree}, \code{neighbor_types}.
}
\description{
Get node metrics
}
\examples{
\donttest{
get_node_metrics("H:UNITCM_H001")
}
}
