% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_ontology_ancestors}
\alias{get_ontology_ancestors}
\title{Get ancestors of an ontology entity}
\usage{
get_ontology_ancestors(tcm_id)
}
\arguments{
\item{tcm_id}{The TCM ontology ID.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns: \code{tcm_id}, \code{name}, \code{name_cn},
\code{level}.
}
\description{
Get ancestors of an ontology entity
}
\examples{
\donttest{
get_ontology_ancestors("TCM_0001")
}
}
