% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{get_ontology_descendants}
\alias{get_ontology_descendants}
\title{Get all descendants of an ontology entity}
\usage{
get_ontology_descendants(tcm_id, max_level = NULL)
}
\arguments{
\item{tcm_id}{The TCM ontology ID.}

\item{max_level}{Maximum depth to descend (integer or \code{NULL} for all).}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Get all descendants of an ontology entity
}
\examples{
\donttest{
get_ontology_descendants("TCM_0001", max_level = 2)
}
}
