% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_similar_datasets}
\alias{get_similar_datasets}
\title{Get similar datasets}
\usage{
get_similar_datasets(submission_id)
}
\arguments{
\item{submission_id}{The submission ID.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns: \code{submission_id},
\code{project_title}, \code{TCM_classification}, \code{similarity_score}.
}
\description{
Find datasets similar to a given submission based on content similarity.
}
\examples{
\donttest{
get_similar_datasets("TMA2025001")
}
}
