% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{list_ontology_categories}
\alias{list_ontology_categories}
\title{List top-level ontology categories}
\usage{
list_ontology_categories()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of level-1 entities.
}
\description{
List top-level ontology categories
}
\examples{
\dontrun{
list_ontology_categories()
}
}
