% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease-formula.R
\name{list_origin_sources}
\alias{list_origin_sources}
\title{List origin sources}
\usage{
list_origin_sources()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{value}, \code{label}, \code{count}.
}
\description{
Returns the top 50 formula origin sources by frequency.
}
\examples{
\dontrun{
list_origin_sources()
}
}
