% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm-term.R
\name{list_term_categories}
\alias{list_term_categories}
\title{List term categories}
\usage{
list_term_categories()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{value}, \code{label}, \code{count}.
}
\description{
List term categories
}
\examples{
\dontrun{
list_term_categories()
}
}
