% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm-term.R
\name{list_term_sources}
\alias{list_term_sources}
\title{List term sources}
\usage{
list_term_sources()
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{value}, \code{label}, \code{count}.
}
\description{
List term sources
}
\examples{
\donttest{
list_term_sources()
}
}
