% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitcm-package.R
\docType{package}
\name{unitcm-package}
\alias{unitcm}
\alias{unitcm-package}
\title{unitcm: R Client for the UniTCM Traditional Chinese Medicine Platform}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to query the UniTCM API \url{https://unitcm.qfxulab.com/api/v1}, covering herb exploration, compound/ADMET data, disease-formula associations, TCM ontology, transcriptomics, and gene-disease analysis (MIDAS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zx122ty.github.io/UniTCM_R_Package/}
  \item \url{https://github.com/zx122ty/UniTCM_R_Package}
  \item Report bugs at \url{https://github.com/zx122ty/UniTCM_R_Package/issues}
}

}
\author{
\strong{Maintainer}: Xiao Zheng \email{12519088@zju.edu.cn} (\href{https://orcid.org/0009-0000-3918-5725}{ORCID})

Authors:
\itemize{
  \item Xiao Zheng \email{12519088@zju.edu.cn} (\href{https://orcid.org/0009-0000-3918-5725}{ORCID})
}

}
\keyword{internal}
